move chat interface files
