+++ resources/assets/lib/chat/chat-api.ts
@@ -1,9 +1,9 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-import ChannelJson from 'interfaces/channel-json';
-import ChatUpdatesJson from 'interfaces/chat-updates-json';
-import MessageJson from 'interfaces/message-json';
+import ChannelJson from 'interfaces/chat/channel-json';
+import ChatUpdatesJson from 'interfaces/chat/chat-updates-json';
+import MessageJson from 'interfaces/chat/message-json';
 import { route } from 'laroute';
 import Message from 'models/chat/message';
