+++ resources/assets/lib/interfaces/chat/chat-updates-json.ts
@@ -1,8 +1,8 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-import ChannelJson from 'interfaces/channel-json';
-import MessageJson from 'interfaces/message-json';
+import ChannelJson from 'interfaces/chat/channel-json';
+import MessageJson from 'interfaces/chat/message-json';
 
 export default interface ChatUpdatesJson {
   messages: MessageJson[];
