+++ resources/assets/lib/interfaces/chat/message-json.ts
