+++ resources/assets/lib/models/chat/message.ts
@@ -1,7 +1,7 @@
 // Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the GNU Affero General Public License v3.0.
 // See the LICENCE file in the repository root for full licence text.
 
-import MessageJson from 'interfaces/message-json';
+import MessageJson from 'interfaces/chat/message-json';
 import { escape } from 'lodash';
 import { action, computed, makeObservable, observable } from 'mobx';
 import User from 'models/user';
