+++ resources/assets/lib/stores/channel-store.ts
@@ -8,9 +8,9 @@ import { ChatNewConversationAdded } from 'actions/chat-new-conversation-added';
 import DispatcherAction from 'actions/dispatcher-action';
 import { dispatch, dispatchListener } from 'app-dispatcher';
 import ChatAPI from 'chat/chat-api';
-import ChannelJson, { ChannelType } from 'interfaces/channel-json';
-import ChatUpdatesJson from 'interfaces/chat-updates-json';
-import MessageJson from 'interfaces/message-json';
+import ChannelJson, { ChannelType } from 'interfaces/chat/channel-json';
+import ChatUpdatesJson from 'interfaces/chat/chat-updates-json';
+import MessageJson from 'interfaces/chat/message-json';
 import { groupBy, maxBy } from 'lodash';
 import { action, computed, makeObservable, observable, runInAction } from 'mobx';
 import Channel from 'models/chat/channel';
